refactor evals into their own dir
