+++ backend/evals/evals.py
@@ -1,13 +1,14 @@
-from pathlib import Path
-from typing import Tuple
 import modal
 import time
-from common import app
-from generator import Model
-from eval_aesthetics import ImprovedAestheticPredictor
-from eval_scannability import detect_qr_ensemble
 
-runner_image = modal.Image.debian_slim().pip_install("qrcode", "wonderwords", "Pillow")
+# Import our app and Generation model
+from .common import app, generate_image
+
+# Import our chosen eval functions
+from .aesthetics import ImprovedAestheticPredictor
+from .scannability import detect_qr_ensemble
+
+runner_image = modal.Image.debian_slim().pip_install("wonderwords")
 
 URLS = [
     "https://www.google.com",
@@ -37,8 +38,8 @@ RANDOM_SEED = 42
 
 AESTHETICS_THRESHOLD = 6.0
 
-N_TESTS = 50
-K_SAMPLES = 10
+N_TESTS = 1
+N_SAMPLES = 3
 
 MINUTE = 60
 
@@ -53,6 +54,8 @@ def estimate_pass_at_k(n: int, c: int, k: int) -> float:
         c: Number of correct samples
         k: Number of trials
     """
+    if k > n and c == 0:
+        return 0.0
     if n - c < k:
         return 1.0
     return 1.0 - np.prod(1.0 - k / np.arange(n - c + 1, n + 1))
@@ -63,8 +66,10 @@ def run_aesthetics_eval(generated_images):
     import numpy as np
     predictor = ImprovedAestheticPredictor()
 
-    # split into sets of K_SAMPLES
-    tests = [generated_images[i:i+K_SAMPLES] for i in range(0, len(generated_images), K_SAMPLES)]
+    # split into sets of N_SAMPLES
+    tests = [generated_images[i:i+N_SAMPLES] 
+        for i in range(0, len(generated_images), N_SAMPLES)
+    ]
 
     results = []
     for test_idx, test in enumerate(tests, 1):
@@ -73,12 +78,12 @@ def run_aesthetics_eval(generated_images):
         num_correct = sum(passed)
         
         # Calculate various pass@k estimates
-        pass_at_1 = estimate_pass_at_k(K_SAMPLES, num_correct, 1)
-        pass_at_3 = estimate_pass_at_k(K_SAMPLES, num_correct, 3)
-        pass_at_10 = estimate_pass_at_k(K_SAMPLES, num_correct, 10)
+        pass_at_1 = estimate_pass_at_k(N_SAMPLES, num_correct, 1)
+        pass_at_3 = estimate_pass_at_k(N_SAMPLES, num_correct, 3)
+        pass_at_10 = estimate_pass_at_k(N_SAMPLES, num_correct, 10)
         
         results.append({
-            'num_samples': K_SAMPLES,
+            'num_samples': N_SAMPLES,
             'num_correct': num_correct,
             'pass@1': pass_at_1,
             'pass@3': pass_at_3,
@@ -87,7 +92,7 @@ def run_aesthetics_eval(generated_images):
         
         print(f"\nAesthetics Test {test_idx}:")
         print(f"Passed: {passed}")
-        print(f"Samples passed: {num_correct}/{K_SAMPLES}")
+        print(f"Samples passed: {num_correct}/{N_SAMPLES}")
         print(f"Estimated pass@1: {pass_at_1:.2%}")
         print(f"Estimated pass@3: {pass_at_3:.2%}")
         print(f"Estimated pass@10: {pass_at_10:.2%}")
@@ -110,8 +115,10 @@ def run_aesthetics_eval(generated_images):
 def run_scannability_eval(generated_images):
     import numpy as np
 
-    # split into sets of K_SAMPLES
-    tests = [generated_images[i:i+K_SAMPLES] for i in range(0, len(generated_images), K_SAMPLES)]
+    # split into sets of N_SAMPLES
+    tests = [generated_images[i:i+N_SAMPLES] 
+        for i in range(0, len(generated_images), N_SAMPLES)
+    ]
 
     results = []
     for test_idx, test in enumerate(tests, 1):
@@ -120,12 +127,12 @@ def run_scannability_eval(generated_images):
         num_correct = sum(passed)
 
         # Calculate various pass@k estimates
-        pass_at_1 = estimate_pass_at_k(K_SAMPLES, num_correct, 1)
-        pass_at_3 = estimate_pass_at_k(K_SAMPLES, num_correct, 3)
-        pass_at_10 = estimate_pass_at_k(K_SAMPLES, num_correct, 10)
+        pass_at_1 = estimate_pass_at_k(N_SAMPLES, num_correct, 1)
+        pass_at_3 = estimate_pass_at_k(N_SAMPLES, num_correct, 3)
+        pass_at_10 = estimate_pass_at_k(N_SAMPLES, num_correct, 10)
 
         results.append({
-            'num_samples': K_SAMPLES,
+            'num_samples': N_SAMPLES,
             'num_correct': num_correct,
             'pass@1': pass_at_1,
             'pass@3': pass_at_3,
@@ -134,7 +141,7 @@ def run_scannability_eval(generated_images):
 
         print(f"\nScannability Test {test_idx}:")
         print(f"Passed: {passed}")
-        print(f"Samples passed: {num_correct}/{K_SAMPLES}")
+        print(f"Samples passed: {num_correct}/{N_SAMPLES}")
         print(f"Estimated pass@1: {pass_at_1:.2%}")
         print(f"Estimated pass@3: {pass_at_3:.2%}")
         print(f"Estimated pass@10: {pass_at_10:.2%}")
@@ -161,9 +168,6 @@ def run_evals():
     '''
     Modal function to run all evals
     '''
-    import base64
-    import io
-    import qrcode
     import random
     from wonderwords import RandomWord
     random.seed(RANDOM_SEED)
@@ -172,11 +176,6 @@ def run_evals():
     for i in range(N_TESTS):
         # generate vanilla qr code
         qr_url = URLS[random.randint(0, len(URLS) - 1)]
-        image = qrcode.make(qr_url) 
-        buffer = io.BytesIO()
-        image.save(buffer, format="PNG")
-        image_bytes = buffer.getvalue()
-        image_base64 = base64.b64encode(image_bytes).decode("utf-8")
 
         # select prompt
         r_gen = RandomWord()
@@ -184,14 +183,14 @@ def run_evals():
         noun = r_gen.word(include_parts_of_speech=["noun"])
         prompt = PROMPT_TEMPLATE.format(adjective=adjective, noun=noun)
 
-        for _ in range(K_SAMPLES):
+        for _ in range(N_SAMPLES):
             # generate multiple images for the same prompt+qr, for pass@k calculation
-            starmap_args.append((prompt, image_base64))
+            starmap_args.append((prompt, qr_url))
 
     # generate images in parallel
     images_bytes = []
     i = 0
-    for image_bytes in Model.generate.starmap(starmap_args):
+    for image_bytes in generate_image.starmap(starmap_args):
         print(f"Generated image {i+1} / {len(starmap_args)}")
         images_bytes.append(image_bytes)
         i += 1
@@ -219,7 +218,7 @@ def evals():
 
     wandb_run.summary.update({
         "n_tests": N_TESTS,
-        "k_samples": K_SAMPLES,
+        "n_samples": N_SAMPLES,
         "aesthetics": aesthetics_score,
         "scannability": scannability_score,
     })
@@ -228,7 +227,7 @@ def evals():
 
     print("Evals complete!")
     print("N_TESTS:", N_TESTS)
-    print("S_SAMPLES:", K_SAMPLES)
+    print("N_SAMPLES:", N_SAMPLES)
     print("\nAesthetics:")
     for k, v in aesthetics_score.items():
         print(f"{k}: {v:.2%}")
