+++ backend/generator/__init__.py
@@ -1,0 +1,4 @@
+from .common import app
+from .generator import Model
+
+__all__ = ["app", "Model"]
